backend/requirements.txt
Django==5.0.4
djangorestframework==3.15.2
psycopg2-binary==2.9.9
python-decouple==3.8
django-cors-headers==4.3.1
drf-spectacular==0.27.2
djangorestframework-simplejwt==5.3.1
